# perf: Optimize negative sampling with vectorized implementation