## Summary

This new implementation leverages vectorization to generate a batch of candidates and filters them against pre-computed interaction sets, greatly speeding up the process.

Resolves #7 

## Key changes:

- Introduce `training_set_uid`, a NumPy array storing sets of interacted items for each user, enabling O(1) average time complexity for lookups.
- Replace the iterative negative sampling with a vectorized strategy to reduce expensive loop overhead.
- Standardize on NumPy's random functions (e.g., `random.shuffle` -> `np.random.shuffle`) for consistency and performance.

## Results

The new implementation has been verified to be functionally correct. The minor fluctuations observed in the evaluation metrics below are within the expected range and are attributed to the inherent randomness of the model's training process.
The performance metrics remain consistent, confirming the validity of the optimization.

### Correctness

This is the new version:

> Top-10 Overall Test Performance:
> Hit@10: 0.0556±0.0000, Precision@10: 0.1743±0.0000, Recall@10: 0.0776±0.0000, NDCG@10: 0.1921±0.0000
> Top-10 Cold-Start Test Performance:
> Hit@10: 0.0901±0.0000, Precision@10: 0.1550±0.0000, Recall@10: 0.1200±0.0000, NDCG@10: 0.1868±0.0000
> Top-10 Warm-Start Test Performance:
> Hit@10: 0.1047±0.0000, Precision@10: 0.1431±0.0000, Recall@10: 0.1416±0.0000, NDCG@10: 0.1845±0.0000
> ********************************************************************************
> Top-20 Overall Test Performance:
> Hit@20: 0.0960±0.0000, Precision@20: 0.1504±0.0000, Recall@20: 0.1290±0.0000, NDCG@20: 0.1889±0.0000
> Top-20 Cold-Start Test Performance:
> Hit@20: 0.1555±0.0000, Precision@20: 0.1338±0.0000, Recall@20: 0.1998±0.0000, NDCG@20: 0.2002±0.0000
> Top-20 Warm-Start Test Performance:
> Hit@20: 0.1716±0.0000, Precision@20: 0.1172±0.0000, Recall@20: 0.2235±0.0000, NDCG@20: 0.2002±0.0000


This is the old version:

********************************************************************************
Top-10 Overall Test Performance:
Hit@10: 0.0556±0.0000, Precision@10: 0.1741±0.0000, Recall@10: 0.0774±0.0000, NDCG@10: 0.1921±0.0000
Top-10 Cold-Start Test Performance:
Hit@10: 0.0886±0.0000, Precision@10: 0.1523±0.0000, Recall@10: 0.1156±0.0000, NDCG@10: 0.1835±0.0000
Top-10 Warm-Start Test Performance:
Hit@10: 0.1061±0.0000, Precision@10: 0.1451±0.0000, Recall@10: 0.1439±0.0000, NDCG@10: 0.1881±0.0000
********************************************************************************
Top-20 Overall Test Performance:
Hit@20: 0.0956±0.0000, Precision@20: 0.1498±0.0000, Recall@20: 0.1271±0.0000, NDCG@20: 0.1881±0.0000
Top-20 Cold-Start Test Performance:
Hit@20: 0.1526±0.0000, Precision@20: 0.1312±0.0000, Recall@20: 0.1925±0.0000, NDCG@20: 0.1958±0.0000
Top-20 Warm-Start Test Performance:
Hit@20: 0.1745±0.0000, Precision@20: 0.1192±0.0000, Recall@20: 0.2266±0.0000, NDCG@20: 0.2043±0.0000
********************************************************************************

### Performance Benchmark

The benchmark was run on an i5-13600KF & RTX 4090 setup.

For GAR on Movielens:

Old Version Time: 1.2487±0.0000 seconds per epoch.

New Version Time: 0.9329±0.0000 seconds per epoch.